refactor: 타입정리 / material supabase 데이터 구조 수정
