+++ package.json
@@ -18,9 +18,8 @@
     "@svgr/webpack": "^8.1.0",
     "@tanstack/react-query": "^5.50.1",
     "@tanstack/react-query-devtools": "^5.50.1",
-    "axios": "^1.7.2",
-    "validator": "^13.12.0",
     "@types/dompurify": "^3.0.5",
+    "axios": "^1.7.2",
     "class-variance-authority": "^0.7.0",
     "dayjs": "^1.11.11",
     "dompurify": "^3.1.6",
@@ -31,6 +30,7 @@
     "sweetalert2": "^11.12.2",
     "swiper": "^11.1.4",
     "uuid": "^10.0.0",
+    "validator": "^13.12.0",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
